Add model for MovieSearch multi
Codable model for /search/multi api call added as "MovieSearchMilti"
