+++ NetflixClone/Classes/Models/MovieSearchMulti.swift
@@ -1,0 +1,48 @@
+//
+//  MovieSearchMulti.swift
+//  NetflixClone
+//
+//  Created by Joseph Ching on 2022-09-27.
+//
+
+import Foundation
+
+struct MovieSearchMulti: Codable, Hashable {
+    let id: Int
+    let backdropPath: String?
+    let firstAirDate: String?
+    let genreIDS: [Int]
+    let mediaType: String?
+    let name: String?
+    let originCountry: [String]?
+    let originalLanguage: String?
+    let originalName: String?
+    let overview: String?
+    let popularity: Double?
+    let posterPath: String?
+    let voteAverage: Double?
+    let voteCount: Int?
+    let adult: Bool?
+    let originalTitle, releaseDate, title: String?
+    let video: Bool?
+
+    enum CodingKeys: String, CodingKey {
+        case backdropPath = "backdrop_path"
+        case firstAirDate = "first_air_date"
+        case genreIDS = "genre_ids"
+        case id
+        case mediaType = "media_type"
+        case name
+        case originCountry = "origin_country"
+        case originalLanguage = "original_language"
+        case originalName = "original_name"
+        case overview, popularity
+        case posterPath = "poster_path"
+        case voteAverage = "vote_average"
+        case voteCount = "vote_count"
+        case adult
+        case originalTitle = "original_title"
+        case releaseDate = "release_date"
+        case title, video
+    }
+}
