+++ NetflixClone/Classes/Networking/NetworkingManager.swift
@@ -10,9 +10,7 @@ import Foundation
 class NetworkingManager {
     static let shared = NetworkingManager()
     private let session = URLSession.shared
-    
     private init() { }
-    
     func responseData<T: Decodable>(_ request: URLRequest) async throws -> T{
         return try await withCheckedThrowingContinuation { continuation in
             self.session.dataTask(with: request) { data, response, error in
@@ -31,113 +29,5 @@ class NetworkingManager {
             }.resume()
         }
     }
-}
-
-// MARK: - Movie
-struct Movie: Codable, Hashable {
-    let id: Int
-    let adult: Bool?
-    let backdropPath: String?
-    let genreIDS: [Int]?
-    let originalLanguage, originalTitle, overview: String?
-    let popularity: Double?
-    let posterPath, releaseDate, title: String?
-    let video: Bool?
-    let voteAverage: Double?
-    let voteCount: Int?
-
-    enum CodingKeys: String, CodingKey {
-        case adult
-        case backdropPath = "backdrop_path"
-        case genreIDS = "genre_ids"
-        case id
-        case originalLanguage = "original_language"
-        case originalTitle = "original_title"
-        case overview, popularity
-        case posterPath = "poster_path"
-        case releaseDate = "release_date"
-        case title, video
-        case voteAverage = "vote_average"
-        case voteCount = "vote_count"
-    }
-}
-
-
-struct Item: Sendable, Hashable {
-    let name: String
-    let idNumber: Int
-    let rating: Double
-    let tags: [String]
-}
-
-
-
-// MARK: Actor
-struct Actor: Codable {
-    let adult: Bool
-    let gender, id: Int
-    let knownFor: [KnownFor]
-    let knownForDepartment: String?
-    let name: String
-    let popularity: Double
-    let profilePath: String
-
-    enum CodingKeys: String, CodingKey {
-        case adult, gender, id
-        case knownFor = "known_for"
-        case knownForDepartment = "known_for_department"
-        case name, popularity
-        case profilePath = "profile_path"
-    }
-}
-
-// MARK: - KnownFor
-struct KnownFor: Codable {
-    let adult: Bool?
-    let backdropPath: String?
-    let genreIDS: [Int]
-    let id: Int
-    let mediaType: MediaType
-    let originalLanguage: OriginalLanguage
-    let originalTitle: String?
-    let overview, posterPath: String
-    let releaseDate, title: String?
-    let video: Bool?
-    let voteAverage: Double
-    let voteCount: Int
-    let firstAirDate, name: String?
-    let originCountry: [String]?
-    let originalName: String?
-
-    enum CodingKeys: String, CodingKey {
-        case adult
-        case backdropPath = "backdrop_path"
-        case genreIDS = "genre_ids"
-        case id
-        case mediaType = "media_type"
-        case originalLanguage = "original_language"
-        case originalTitle = "original_title"
-        case overview
-        case posterPath = "poster_path"
-        case releaseDate = "release_date"
-        case title, video
-        case voteAverage = "vote_average"
-        case voteCount = "vote_count"
-        case firstAirDate = "first_air_date"
-        case name
-        case originCountry = "origin_country"
-        case originalName = "original_name"
-    }
-}
-
-enum MediaType: String, Codable {
-    case movie = "movie"
-    case tv = "tv"
-}
-
-enum OriginalLanguage: String, Codable {
-    case en = "en"
-    case ko = "ko"
-    case tr = "tr"
 }
