+++ NetflixClone/Classes/Scenes/Search/SearchViewController.swift
@@ -1,48 +1,25 @@
-////
-////  SearchViewController.swift
-////  NetflixClone
-////
-////  Created by Joseph Ching on 2022-09-11.
-////
 //
-//import UIKit
+//  SearchViewController.swift
+//  NetflixClone
 //
-//class SearchViewController: UIViewController {
+//  Created by Joseph Ching on 2022-09-11.
 //
-//    @IBOutlet weak var searchCollectionView: UICollectionView!
-//
-//    override func viewDidLoad() {
-//        super.viewDidLoad()
-//        searchCollectionView.dataSource = self
-//        searchCollectionView.delegate = self
-//
-//        let nibName = "\(SearchCollectionViewCell.self)"
-//        searchCollectionView.register(UINib(nibName: nibName, bundle: nil), forCellWithReuseIdentifier: nibName)
-//    }
-//
-//}
-//
-//extension SearchViewController: UICollectionViewDataSource, UICollectionViewDelegate {
-//    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-//        return 10
-//    }
-//
-//    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-//        let collectionViewCell = collectionView.dequeueReusableCell(withReuseIdentifier: "\(SearchCollectionViewCell.self)", for: indexPath) as! SearchCollectionViewCell
-//        return collectionViewCell
-//    }
-//
-//}
-//
-//extension SearchViewController: UICollectionViewDelegateFlowLayout {
-//    func collectionView(
-//        _ collectionView: UICollectionView,
-//        layout collectionViewLayout: UICollectionViewLayout,
-//        sizeForItemAt indexPath: IndexPath
-//    ) -> CGSize
-//    {
-//        let width = collectionView.bounds.width/3
-//        let height = width * 1.5
-//        return CGSize(width: width, height: height)
-//    }
-//}
+
+import UIKit
+
+class SearchViewController: UIViewController {
+    // MARK: Properties
+    private enum Defaults {
+        static let itemsPerPage = 10
+    }
+    var collectionView: UICollectionView! = nil
+    private let remote = DiscoverMoviesRemote()
+//    var dataSource: UICollectionViewDiffableDataSource<Section, Movie>! = nil
+//    var currentSnapshot: NSDiffableDataSourceSnapshot<Section, Movie>! = nil
+    private var currentPage = 1
+    override func viewDidLoad() {
+        super.viewDidLoad()
+         
+    }
+
+}
