+++ NetflixClone/Classes/Scenes/TabBar/TabBarController.swift
@@ -13,21 +13,19 @@ class TabBarController: UITabBarController {
         super.viewDidLoad()
 
         let homeViewController = makeViewController(controller: HomeViewController(), title: "Home", systemName: "house", tag: 0)
-//        let searchViewController = makeViewController(controller: SearchViewController(), title: "Search", systemName: "magnifyingglass", tag: 1)
+        let searchViewController = makeViewController(controller: SearchViewController(), title: "Search", systemName: "magnifyingglass", tag: 1)
         let favouritesVieController = makeViewController(controller: FavouritesViewController(), title: "Favourites", systemName: "star", tag: 2)
         let options = makeViewController(controller: MoreViewController(), title: "More", systemName: "square.and.arrow.up", tag: 3)
-//        viewControllers = [homeViewController, searchViewController, favouritesVieController, options]
+        viewControllers = [homeViewController, searchViewController, favouritesVieController, options]
         viewControllers = [homeViewController, favouritesVieController, options]
 
     }
-    
     func makeViewController(
         controller: UIViewController,
         title: String,
-        systemName : String,
+        systemName: String,
         tag: Int
-    ) -> UINavigationController
-    {
+    ) -> UINavigationController {
         controller.tabBarItem = UITabBarItem(
             title: title,
             image: UIImage(systemName: systemName),
